finish save label and time arange, use mianno.anno_length and midata.data_duration to control the total seconds.

+++ MiSleep/utils/annotation.py
@@ -39,7 +39,7 @@ def start_end2mianno(start_end):
     """Transfer start_end to [[1, 20, 'spindle'], [30, 50, 'SWA'], ...]"""
     if start_end != [] or start_end is not None:
         start_end = [each.split(', ') for each in start_end]
-        start_end = [[int(each[1]), int(each[4]), each[7]] for each in start_end]
+        start_end = [[float(each[1]), float(each[4]), each[7]] for each in start_end]
         return start_end
     return []
 
@@ -47,4 +47,5 @@ def start_end2mianno(start_end):
 def sleep_state2mianno(sleep_state):
     """Transfer sleep_state to [1, 1, 1, 1, 1, 1, 1, 2, 2, 2, 2, ...]"""
     start_end = [each.split(', ') for each in sleep_state]
+
     return [item for each in start_end for item in [int(each[6])] * (int(each[4]) - int(each[1]) + 1)]
